pyproject.toml
[project]
name = "knowledge-base"
version = "0.1.0"
description = ""
authors = [
    {name = "RandomCyberCoder"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pypdf2 (>=3.0.1,<4.0.0)",
    "pycryptodome (>=3.23.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "qdrant-client (>=1.15.1,<2.0.0)",
    "sentence-transformers (>=5.1.0,<6.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "nltk (>=3.9.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
